+++ bmc-database/src/main/java/com/oracle/bmc/database/model/AutonomousContainerDatabaseSummary.java
@@ -361,12 +361,12 @@ public class AutonomousContainerDatabaseSummary {
     String lifecycleDetails;
 
     /**
-     * The date and time the Autonomous was created.
+     * The date and time the Autonomous Container Database was created.
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("timeCreated")
     java.util.Date timeCreated;
     /**
-     * Database Patch model preference.
+     * Database patch model preference.
      **/
     @lombok.extern.slf4j.Slf4j
     public enum PatchModel {
@@ -412,7 +412,7 @@ public class AutonomousContainerDatabaseSummary {
         }
     };
     /**
-     * Database Patch model preference.
+     * Database patch model preference.
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("patchModel")
     PatchModel patchModel;