+++ bmc-database/src/main/java/com/oracle/bmc/database/model/BackupDestinationDetails.java
@@ -60,32 +60,12 @@ public class BackupDestinationDetails {
             return this;
         }
 
-        @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
-        private java.util.Map<String, String> freeformTags;
-
-        public Builder freeformTags(java.util.Map<String, String> freeformTags) {
-            this.freeformTags = freeformTags;
-            this.__explicitlySet__.add("freeformTags");
-            return this;
-        }
-
-        @com.fasterxml.jackson.annotation.JsonProperty("definedTags")
-        private java.util.Map<String, java.util.Map<String, Object>> definedTags;
-
-        public Builder definedTags(
-                java.util.Map<String, java.util.Map<String, Object>> definedTags) {
-            this.definedTags = definedTags;
-            this.__explicitlySet__.add("definedTags");
-            return this;
-        }
-
         @com.fasterxml.jackson.annotation.JsonIgnore
         private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
 
         public BackupDestinationDetails build() {
             BackupDestinationDetails __instance__ =
-                    new BackupDestinationDetails(
-                            type, id, vpcUser, vpcPassword, freeformTags, definedTags);
+                    new BackupDestinationDetails(type, id, vpcUser, vpcPassword);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
             return __instance__;
         }
@@ -96,9 +76,7 @@ public class BackupDestinationDetails {
                     type(o.getType())
                             .id(o.getId())
                             .vpcUser(o.getVpcUser())
-                            .vpcPassword(o.getVpcPassword())
-                            .freeformTags(o.getFreeformTags())
-                            .definedTags(o.getDefinedTags());
+                            .vpcPassword(o.getVpcPassword());
 
             copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
             return copiedBuilder;
@@ -182,24 +160,6 @@ public class BackupDestinationDetails {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("vpcPassword")
     String vpcPassword;
-
-    /**
-     * Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace.
-     * For more information, see [Resource Tags](https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
-     * <p>
-     * Example: `{\"Department\": \"Finance\"}`
-     *
-     **/
-    @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
-    java.util.Map<String, String> freeformTags;
-
-    /**
-     * Defined tags for this resource. Each key is predefined and scoped to a namespace.
-     * For more information, see [Resource Tags](https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
-     *
-     **/
-    @com.fasterxml.jackson.annotation.JsonProperty("definedTags")
-    java.util.Map<String, java.util.Map<String, Object>> definedTags;
 
     @com.fasterxml.jackson.annotation.JsonIgnore
     private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();