+++ bmc-database/src/main/java/com/oracle/bmc/database/model/MaintenanceRunSummary.java
@@ -227,6 +227,7 @@ public class MaintenanceRunSummary {
         Succeeded("SUCCEEDED"),
         Skipped("SKIPPED"),
         Failed("FAILED"),
+        Updating("UPDATING"),
 
         /**
          * This value is used if a service returns a value for this enum that is not recognized by this
@@ -302,6 +303,7 @@ public class MaintenanceRunSummary {
     public enum TargetResourceType {
         AutonomousExadataInfrastructure("AUTONOMOUS_EXADATA_INFRASTRUCTURE"),
         AutonomousContainerDatabase("AUTONOMOUS_CONTAINER_DATABASE"),
+        ExadataDbSystem("EXADATA_DB_SYSTEM"),
 
         /**
          * This value is used if a service returns a value for this enum that is not recognized by this