+++ bmc-filestorage/src/main/java/com/oracle/bmc/filestorage/model/MountTargetSummary.java
@@ -97,6 +97,15 @@ public class MountTargetSummary {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("nsgIds")
+        private java.util.List<String> nsgIds;
+
+        public Builder nsgIds(java.util.List<String> nsgIds) {
+            this.nsgIds = nsgIds;
+            this.__explicitlySet__.add("nsgIds");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("timeCreated")
         private java.util.Date timeCreated;
 
@@ -139,6 +148,7 @@ public class MountTargetSummary {
                             lifecycleState,
                             privateIpIds,
                             subnetId,
+                            nsgIds,
                             timeCreated,
                             freeformTags,
                             definedTags);
@@ -157,6 +167,7 @@ public class MountTargetSummary {
                             .lifecycleState(o.getLifecycleState())
                             .privateIpIds(o.getPrivateIpIds())
                             .subnetId(o.getSubnetId())
+                            .nsgIds(o.getNsgIds())
                             .timeCreated(o.getTimeCreated())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags());
@@ -279,6 +290,16 @@ public class MountTargetSummary {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("subnetId")
     String subnetId;
+
+    /**
+     * A list of Network Security Group [OCIDs](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) associated with this mount target.
+     * A maximum of 5 is allowed.
+     * Setting this to an empty array after the list is created removes the mount target from all NSGs.
+     * For more information about NSGs, see [Security Rules](https://docs.cloud.oracle.com/Content/Network/Concepts/securityrules.htm).
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("nsgIds")
+    java.util.List<String> nsgIds;
 
     /**
      * The date and time the mount target was created, expressed