+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
+## 1.10.0 - 2019-11-26
+### Added
+- Support for maintenance windows on autonomous databases in the Database service
+- Support for getting the compute units (OCPUs) of an Exadata autonomous transaction processing - dedicated resource in the Database service
+
+### Breaking changes
+- Create database home from VM_CLUSTER_BACKUP is removed from Database Service
+
 ## 1.9.5 - 2019-11-19
 ### Added
 - Support for four-byte autonomous system numbers (ASNs) on FastConnect resources in the Networking service